+++ src/matvar_cell.c
@@ -48,7 +48,7 @@ Mat_VarGetCell(matvar_t *matvar,int index)
 
     SafeMulDims(matvar, &nelems);
 
-    if ( 0 <= index && index < nelems )
+    if ( 0 <= index && (size_t)index < nelems )
         cell = *((matvar_t **)matvar->data + index);
 
     return cell;
@@ -172,9 +172,9 @@ Mat_VarSetCell(matvar_t *matvar,int index,matvar_t *cell)
     if ( matvar == NULL || matvar->rank < 1 )
         return NULL;
 
-   SafeMulDims(matvar, &nelems);
+    SafeMulDims(matvar, &nelems);
     cells = (matvar_t**)matvar->data;
-    if ( 0 <= index && index < nelems ) {
+    if ( 0 <= index && (size_t)index < nelems ) {
         old_cell = cells[index];
         cells[index] = cell;
     }
